+++ Examples/NativeTemplatesExample/NativeTemplatesExample/main.m
@@ -1,0 +1,19 @@
+/*
+ * main.m
+ *
+ * This file is a part of the Yandex Advertising Network.
+ *
+ * Version for iOS © 2016 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://legal.yandex.com/partner_ch/
+ */
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}