+++ Examples/NativeTemplatesTableViewExample/NativeTemplatesTableViewExample/NativeBannerTableViewCell.h
@@ -1,0 +1,22 @@
+/*
+ *  NativeBannerTableViewCell.h
+ *
+ * This file is a part of the Yandex Advertising Network.
+ *
+ * Version for iOS © 2016 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://legal.yandex.com/partner_ch/
+ */
+
+#import <UIKit/UIKit.h>
+
+@protocol YMANativeGenericAd;
+@class YMANativeBannerView;
+
+@interface NativeBannerTableViewCell : UITableViewCell
+
+@property (nonatomic, strong) id<YMANativeGenericAd> ad;
+@property (nonatomic, weak, readonly) YMANativeBannerView *adView;
+
+@end