+++ SDK/YandexMobileAds/YMANativeAdTypes.h
@@ -1,0 +1,22 @@
+/*
+ *  YMANativeAdTypes.h
+ *
+ * This file is a part of the Yandex Advertising Network.
+ *
+ * Version for iOS © 2016 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://legal.yandex.com/partner_ch/
+ */
+
+#import <Foundation/Foundation.h>
+
+/**
+ * Content ad type.
+ */
+extern NSString *const kYMANativeAdTypeContent;
+
+/**
+ * App install ad type.
+ */
+extern NSString *const kYMANativeAdTypeAppInstall;